+++ BinanceApiAdapter/Types/BalanceWss.cs
@@ -13,9 +13,9 @@ namespace CryptoSwapMaster.BinanceApiAdapter.Types
         public string Asset { get; set; }
 
         [JsonProperty("f")]
-        public double Free { get; set; }
+        public decimal Free { get; set; }
 
         [JsonProperty("l")]
-        public double Locked { get; set; }
+        public decimal Locked { get; set; }
     }
 }
